CMake: Prevent multiple creation of the targets 'aab' and 'apk'
Suppose you have a project with multiple subdirectories.  In the
subdirectories are find_package calls for Qt, but not in the top-level
file.

Example:
- CMakeLists.txt
- subdir1/CMakeLists.txt
  find_package(Qt5 COMPONENTS Core)
- subdir2/CMakeLists.txt
  find_package(Qt5 COMPONENTS Core)

Both find_package calls are local to their subdirectories, which means
that the second find_package is executed, even though the first was
already successful.  Both calls try to create the targets 'apk' and
'aab'.  This leads to the following CMake error:

    add_custom_target cannot create target "apk" because another target
    with the same name already exists.

Guard the add_custom_target calls such that they do not create a target
if it is already existent.

Fixes: QTBUG-87863
Change-Id: I132b0c7d9db9b2cb33b5e091cf0b216f06cd0d39
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>